refactor: Clean up types
