+++ htmlisp/utilities/parseExpressions.ts
@@ -1,7 +1,6 @@
 import { parseExpression } from "./parseExpression.ts";
 import { applyUtility } from "./applyUtility.ts";
-import type { Attributes } from "./parsers/types.ts";
-import type { Context } from "../types.ts";
+import type { Attributes, Context } from "../types.ts";
 import type { Utilities, Utility } from "../../types.ts";
 
 async function parseExpressions(
