+++ htmlisp/utilities/parsers/types.ts
@@ -1,15 +1,7 @@
-type Attributes = Record<string, string | null>;
-
 type CharacterGenerator = {
   get: (offset?: number) => string | null;
   next: () => string | null;
   previous: () => string | null;
 };
-type Element = {
-  type: string;
-  attributes?: Attributes;
-  children: (string | Element)[];
-  closesWith?: string | null;
-};
 
-export type { Attributes, CharacterGenerator, Element };
+export type { CharacterGenerator };
